refactor: default empty fallback schedule in store via .at()
Co-locate EMPTY_FALLBACK_SCHEDULE with getCurrentFallbackSchedule and use
Array.prototype.at for index lookup.
