+++ apps/antalmanac/src/stores/FallbackStore.ts
@@ -1,12 +1,19 @@
 import type { ShortCourseSchedule } from '@packages/antalmanac-types';
 import { create } from 'zustand';
 
+const EMPTY_FALLBACK_SCHEDULE: ShortCourseSchedule = {
+    scheduleName: '',
+    courses: [],
+    customEvents: [],
+    scheduleNote: '',
+};
+
 interface FallbackStore {
     fallbackMode: boolean;
     fallbackSchedules: ShortCourseSchedule[];
 
     loadFallbackSchedules: (schedules: ShortCourseSchedule[]) => void;
-    getCurrentFallbackSchedule: (currentScheduleIndex: number) => ShortCourseSchedule | undefined;
+    getCurrentFallbackSchedule: (currentScheduleIndex: number) => ShortCourseSchedule;
     getFallbackScheduleNames: () => string[];
 }
 
@@ -19,7 +26,7 @@ export const useFallbackStore = create<FallbackStore>((set, get) => ({
     },
 
     getCurrentFallbackSchedule: (currentScheduleIndex: number) => {
-        return get().fallbackSchedules[currentScheduleIndex];
+        return get().fallbackSchedules.at(currentScheduleIndex) ?? EMPTY_FALLBACK_SCHEDULE;
     },
 
     getFallbackScheduleNames: () => {
